ISPN-4105 Optimize Cluster Listener performance for Replicated Caches
* Changed so REPL doesn't register cluster listener on other nodes
* REPL not notifies during the remote write
