+++ core/src/main/java/org/infinispan/notifications/Listener.java
@@ -205,8 +205,9 @@ public @interface Listener {
 
    /**
     * Specifies whether the event should be fired on the primary data owner of the affected key, or all nodes that see
-    * the update.  In the case of replication, this would be the coordinator.
-    *
+    * the update.
+    * <p>
+    * Note that is value is ignored when {@link org.infinispan.notifications.Listener#clustered()} is true.
     * @return true if the expectation is that only the primary data owner will fire the event, false if all nodes that
     *         see the update fire the event.
     *
