+++ core/src/test/java/org/infinispan/notifications/cachelistener/cluster/ClusterListenerDistTxAddListenerTest.java
@@ -1,0 +1,16 @@
+package org.infinispan.notifications.cachelistener.cluster;
+
+import org.testng.annotations.Test;
+
+/**
+ * Cluster listener test having a configuration of non tx and dist when adding listeners at inopportune times
+ *
+ * @author wburns
+ * @since 7.0
+ */
+@Test(groups = "functional", testName = "notifications.cachelistener.cluster.ClusterListenerDistTxAddListenerTest")
+public class ClusterListenerDistTxAddListenerTest extends AbstractClusterListenerDistAddListenerTest {
+   public ClusterListenerDistTxAddListenerTest() {
+      super(true);
+   }
+}
